WIP: strbuf_realpath(): error out on "b/" if "b" is a regular file
The documentation for strbuf_realpath() specifies that all path
components up to the last dir separator "must denote a valid, existing
directory". However, if we have a path like "/a/b/", where "b" is a
regular file, the function does not report an error. Let's fix that.

In addition, this modification also helps the case where we have
"/a/b/c", and "b" is a regular file. Instead of calling lstat() for both
"/a/b" and "/a/b/c", we error out on the first, saving a little time and
giving a more direct error report.