strbuf_realpath(): error out if non-dir path ends in a slash
The comment above strbuf_realpath() specifies that "everything up to the
last dir_sep must denote a valid, existing directory". However, if we
have a path which ends in a non-directory file followed by a directory
separator (e.g. "/dir/dir2/regular_file/"), the function does not
report an error. Let's fix that.

Note: when the non-directory file is followed by other path components
the function already correctly reports an error. But this patch also
slightly optimizes this case: instead of calling lstat() twice -- for
"/dir/regular_file" and "/dir/regular_file/other" -- we error out on the
first call.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>