WIP: parallel-checkout: auto-enable parallelism on NFS/EFS
Parallel checkout is usually faster than the default sequential version
when writing to NFS mounts. This is true even when the NFS client and
the NFS server are both running on single-core machines, regardless of
the NFS version.

Bellow are the mean runtimes and standard deviations for 5 cold-cache
executions of a local linux-v5.8 clone, with increasing number of
workers for checkout. The NFS client, where Git was invoked, is an
Amazon EC2 c5n.large instance, and the NFS server is a m6g.large
instance. Both have two vCPUs, but only one core was left online.

     nfs 3.0               nfs 4.0               nfs 4.1
1:   183.708 s ± 3.290 s   205.851 s ± 0.844 s   217.317 s ± 3.047 s
2:   130.510 s ± 3.917 s   139.124 s ± 0.772 s   142.963 s ± 0.765 s
4:    89.611 s ± 1.032 s   102.701 s ± 1.665 s    94.728 s ± 1.014 s
8:    68.097 s ± 0.820 s   104.914 s ± 1.239 s    69.359 s ± 0.619 s
16:   63.999 s ± 0.820 s   104.808 s ± 2.279 s    64.843 s ± 0.587 s
32:   62.316 s ± 2.095 s   102.105 s ± 1.537 s    64.122 s ± 0.374 s
64:   63.699 s ± 0.841 s   103.103 s ± 1.319 s    63.532 s ± 0.734 s

The parallel version is also faster on NFS for considerably small
checkouts. Bellow are benchmark numbers for a local clone of the 'bat'
repository, in the same environment of the previous tests. The version
of 'bat' cloned (v0.16.0) has 251 files, which correspond to about 3MB.

     nfs 3.0             nfs 4.0             nfs 4.1
1:   0.853 s ± 0.080 s   0.814 s ± 0.020 s   0.876 s ± 0.065 s
2:   0.671 s ± 0.020 s   0.702 s ± 0.030 s   0.705 s ± 0.030 s
4:   0.530 s ± 0.024 s   0.595 s ± 0.020 s   0.570 s ± 0.030 s
8:   0.470 s ± 0.033 s   0.609 s ± 0.025 s   0.510 s ± 0.031 s
16:  0.469 s ± 0.037 s   0.616 s ± 0.022 s   0.513 s ± 0.030 s
32:  0.487 s ± 0.030 s   0.639 s ± 0.018 s   0.527 s ± 0.028 s
64:  0.520 s ± 0.022 s   0.680 s ± 0.028 s   0.562 s ± 0.026 s

As we can see in the above numbers, we seem to get the best overall
results around 16 workers. Seeking to provide users with the best
performance by default, let's attempt to detect the file system type and
automatically use 16 workers when the checkout path lies in an NFS
mount. If detection fails, we just fall back to the regular sequential
code. Note that `git checkout-index` accepts a --prefix parameter which
can make the checkout path fall outside the repository. This case is
also handled.

Note: Inferring the type of a file system is a platform-specific
problem. For now, the auto-detection is only implemented for linux.
Nevertheless, Git users on other OSes can still benefit from parallel
checkout on NFS by manually setting the `checkout.workers`
configuration. Additionally, for consistency and good usability, the
auto-detection is disabled when the user explicitly set the number of
workers.

Amazon's Elastic File System (EFS), is an elastic NFS implementation. As
we can see in the following benchmark, it might benefit from a greater
number of workers:

     linux clone on efs       bat clone on efs
1:   1419.769 s ± 114.150 s   6.119 s ± 0.365 s
2:    781.145 s ± 13.345 s    4.499 s ± 0.315 s
4:    491.822 s ± 11.456 s    3.604 s ± 0.121 s
8:    347.835 s ± 7.976 s     3.055 s ± 0.199 s
16:   274.520 s ± 9.653 s     2.953 s ± 0.117 s
32:   297.373 s ± 11.527 s    3.011 s ± 0.187 s
64:   261.074 s ± 12.806 s    3.082 s ± 0.211 s

So, if we detect EFS, lets use 64 workers by default.