parallel-checkout: auto-enable parallelism on NFS
Parallel checkout is usually faster than the default sequential version
when writing to NFS mounts. This is true even when the NFS client and
the NFS server are both running on single-core machines, and regardless
of the NFS version.

Below are the mean runtimes and standard deviations for 5 cold-cache
executions of a local linux-v5.8 clone, on Linux, with increasing number
of checkout workers. The test was repeated with different NFS versions
and also on Amazon EFS (using NFS v4.1 and with an upcoming optimization
enabled). The NFS client, where Git was invoked, is an Amazon EC2
c5n.large instance, and the NFS server is a m6g.large instance. Both
have two vCPUs, but I disabled the additional core, to simulate
single-core machines.

    nfs 3.0              nfs 4.0              nfs 4.1              EFS
1:  183.708 s ± 3.290 s  205.851 s ± 0.844 s  217.317 s ± 3.047 s  936.437 s ± 5.735 s
2:  130.510 s ± 3.917 s  139.124 s ± 0.772 s  142.963 s ± 0.765 s  547.299 s ± 6.000 s
4:   89.611 s ± 1.032 s  102.701 s ± 1.665 s   94.728 s ± 1.014 s  362.454 s ± 7.331 s
8:   68.097 s ± 0.820 s  104.914 s ± 1.239 s   69.359 s ± 0.619 s  266.756 s ± 1.345 s
16:  63.999 s ± 0.820 s  104.808 s ± 2.279 s   64.843 s ± 0.587 s  226.250 s ± 3.021 s
32:  62.316 s ± 2.095 s  102.105 s ± 1.537 s   64.122 s ± 0.374 s  210.673 s ± 6.514 s
64:  63.699 s ± 0.841 s  103.103 s ± 1.319 s   63.532 s ± 0.734 s  209.553 s ± 3.442 s

The parallel version was also faster on small checkouts. Below are the
benchmark numbers for a local clone of the 'bat' repository, in the same
environment of the previous test. The version of 'bat' cloned (v0.16.0)
has 251 files, which correspond to about 3MB.

     nfs 3.0             nfs 4.0             nfs 4.1             EFS
1:   0.853 s ± 0.080 s   0.814 s ± 0.020 s   0.876 s ± 0.065 s   3.959 s ± 0.121 s
2:   0.671 s ± 0.020 s   0.702 s ± 0.030 s   0.705 s ± 0.030 s   2.987 s ± 0.100 s
4:   0.530 s ± 0.024 s   0.595 s ± 0.020 s   0.570 s ± 0.030 s   2.582 s ± 0.173 s
8:   0.470 s ± 0.033 s   0.609 s ± 0.025 s   0.510 s ± 0.031 s   2.368 s ± 0.067 s
16:  0.469 s ± 0.037 s   0.616 s ± 0.022 s   0.513 s ± 0.030 s   2.310 s ± 0.078 s
32:  0.487 s ± 0.030 s   0.639 s ± 0.018 s   0.527 s ± 0.028 s   2.276 s ± 0.115 s
64:  0.520 s ± 0.022 s   0.680 s ± 0.028 s   0.562 s ± 0.026 s   2.340 s ± 0.134 s

Across these NFS versions, we get the best overall numbers with about 16
workers. Seeking to provide users with the best performance by default,
let's attempt to detect the file system type and automatically use 16
workers on NFS, when `checkout.workers` is not set. If detection fails,
we just fall back to the regular sequential code. Note that `git
checkout-index` accepts a --prefix option which can make the checkout
path fall outside the repository. This case is also handled.

Note: inferring the type of a file system is a platform-specific
problem. For now, the auto-detection is only implemented for Linux,
where we also verified that 16 workers normally bring a good
performance.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>