WIP: parallel-checkout: auto-enable parallelism on NFS/EFS
Parallel checkout is usually faster than the default sequential version
when writing to NFS mounts. This is true even when the NFS client and
the NFS server are both running on single-core machines, regardless of
the NFS version:

< Numbers NFS different versions, Linux >

And this is also true for considerably small checkouts:

< Numbers NFS different versions, bat >

Seeking to provide users with the best performance, by default, let's
attempt to detect the file system type and automatically enable
parallelism for NFS. If detection fails, we just fall back to the
regular sequential code.

Note: Inferring the type of a file system is a platform-specific
problem. For now, the auto-detection is only implemented for linux.
Nevertheless, Git users on other OSes can still benefit from parallel
checkout on NFS by manually setting the `checkout.workers`
configuration. Additionally, for consistency and good usability, the
auto-detection is disabled when the user explicitly set the number of
workers.

We will use 16 workers on "regular" NFS, since this seems to bring a
good overall gain in performance across the different versions of the
protocol. The Amazon Elastic File System (EFS), which also communicates
through the NFS protocol, can benefit of an even higher number of
workers:

< Numbers EFS, Linux and bat >

So, in the EFS case, we will use 64 workers by default.