checkout: add guards against trying to write submodules to custom paths
On request, checkout_entry() can write the given entry to a tempfile or
prepend a prefix to its path. However, neither of these options are
compatible with submodules, and the function also doesn't error out when
they are erroneously used. Instead, it silently creates a directory at
the custom path and updates the files at the submodule's directory
itself. Fortunately, it looks like we currently don't have any code
paths that call checkout_entry() with this invalid combination of
parameters. But to future-proof the function, let's add guards to it and
its callee, write_entry().

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>