write_entry(): fix misuses of `path` in error messages
The strings `path` and `ce->name` at write_entry() usually have the same
contents, but that's not the case when using a checkout prefix or
writing to a tempfile. In fact, when writing to a tempfile, `path` will
be either empty or dirty. So these variables cannot be used
interchangeably. In this sense, fix wrong uses of `path` in error
messages where it should really be `ce->name`. There doesn't seem to be
any misuses in the other way around.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>