checkout-index: improve --temp output on mkstemp() failure
When `git checkout-index --temp` fails to generate an unique temporary
filename for all of the to-be-written entries, we end up with an output
like:

error: unable to create file .merge_file_XXXXXX: File exists
error: unable to create file .merge_file_XXXXXX: File exists
...
.merge_file_XXXXXX	file_n
.merge_file_XXXXXX	file_n+1
...

The error message suggests that git failed to create the file
'.merge_file_XXXXXX', when in reality it has run out of temporary
filenames with that template. And the lines associating
'.merge_file_XXXXXX' with index entries are not very useful since the
temporary files weren't really created. Fix those issues using a more
informative error message, and clearing the output `path` parameter when
we fail to generate a tempname, so that these entries are omitted from
the output list.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>