grep: honor sparse-checkout on working tree searches
On a sparse checked out working tree, `git grep` (without --cached) will
search the cache when an entry matches the search pathspec and has the
SKIP_WORKTREE bit set. This is confusing as the output mixes both
working tree and index results, without distinguishing them.
Furthermore, there is no current option to tell `git grep` to combine
searches in the index, the working tree, and/or past revisions; so the
command is not expected to mix results in any case.

Fix that by teaching grep to honor the sparse-checkout rules for working
tree searches. If the user wants to grep paths outside the current
sparse-checkout definition, they may either update the sparsity rules or
use --cached, which will search cache entries with and without the
SKIP_WORKTREE bit set.

Note: it would also be interesting to add a configuration option that
allow users to search working tree files outside the sparsity rules
(e.g. a file that was manually created after a `git sparse-checkout
init`), and to restrict searches in the index and past revisions too.
But these ideas are left as future improvements, to avoid conflicting
with other topics in flight and to focus on the bug fix for now.

Suggested-by: Elijah Newren <newren@gmail.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>