add: warn when pathspec only matches SKIP_WORKTREE entries
`git add` already refrains from updating SKIP_WORKTREE entries, but it
silently succeeds when a pathspec only matches these entries. Instead,
let's warn the user and display a hint on how to update these entries.

Note that the warning is only shown if the pathspec matches no untracked
paths in the working tree and only matches index entries with the
SKIP_WORKTREE bit set. Performance-wise, this patch doesn't change the
number of ce_path_match() calls in the worst case scenario (because we
still need to check the sparse entries for the warning). But in the
general case, it avoids unnecessarily calling this function for each
SKIP_WORKTREE entry.

A warning message was chosen over erroring out right away to reproduce
the same behavior `add` already exhibits with ignored files. This also
allow users to continue their workflow without having to invoke `add`
again with only the matching pathspecs, as the matched files will have
already been added.

Note: refresh_index() was changed to only mark matches with
no-SKIP-WORKTREE entries in the `seen` output parameter. This is exactly
the behavior we want for `add`, and only `add` calls this function with
a non-NULL `seen` pointer. So the change brings no side effect on
other callers.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>