refresh_index(): add REFRESH_DONT_MARK_SPARSE_MATCHES flag
refresh_index() optionally takes a seen[] array to mark the pathspec
items that had matches in the index. This is used by `git add --refresh`
to find out if there was any pathspec without matches, and display an
error accordingly.

In the following patch, `git add` will also learn to warn about
pathspecs that match no eligible path for updating, but do match sparse
entries. For that, we will need a seen[] array marked exclusively with
matches from dense entries. To avoid having to call ce_path_match()
again for these entries after refresh_index() returns, add a flag that
implements this restriction inside the function itself.

Note that refresh_index() does not update sparse entries, regardless of
passing the flag or not. The flag only controls whether matches with
these entries should appear in the seen[] array.

While we are here, also realign the REFRESH_* flags and convert the hex
values to the more natural bit shift format, which makes it easier to
spot holes.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>