refresh_index(): add flag to ignore SKIP_WORKTREE entries
refresh_index() doesn't update SKIP_WORKTREE entries, but it still
matches them against the given pathspecs, and marks the matches on the
seen[] array. In the following patch, one caller will need
refresh_index() to ignore SKIP_WORKTREE entries entirely, so that it
can obtain a seen[] array with only non-SKIP_WORKTREE matches. Add a
flag that implements this behavior.

While we are here, also realign the REFRESH_* flags and convert the hex
values to the more natural bit shift format, which makes it easier to
spot holes.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>