format-patch: warn if commit msg contains a patch break
When applying a patch, there are a few special strings that `git am`
looks for to know where the message ends and the actual diff starts. One
example is '---'. However, if these strings appear in the commit message
itself, `am` might get confused and fail to apply the patch properly.
To help avoid these problems, let's make `git format-patch` warns when
it sees a commit containing one of the said strings.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>