WIP: clean: remove unnecessary variable
The variable `matches` used to be assigned with the return of a
`dir_path_match()` call, which was removed in 95c11ecc73 ("Fix
error-prone fill_directory() API; make it only return matches",
2020-04-01). So now `matches` is unnecessary, and the condition
`matches != MATCHED_EXACTLY` will always evaluate to true.

TODO: is it OK that we now `continue` regardless of the match type?

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>