t2080: fix cp invocation to copy symlinks instead of following them
t2080 makes a few copies of a test repository and later perform a branch
switch on each one of the copies to verify that parallel checkout and
sequential checkout produce the same results. However, the repository is
copied with `cp -R` which, on some systems, defaults to following
symlinks on the directory hierarchy and copying their target files
instead of the symlinks themselves. AIX is one example of system where
this happens. Because the file types are not preserved, the copied
repositories have dirty paths, which leads git to abort the checkout
operation. This makes the test fail on these systems.

Let's fix this by using the POSIX flag '-P' when copying the repository,
which forces cp to copy the symlinks instead of following them. With
this change, t2080 now passes on AIX.

Reported-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>